backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
pydantic==2.11.7
python-multipart==0.0.9
pdfminer.six==20250506
pypdf==5.1.0
fpdf2==2.8.4
